+++ CHANGES.rst
@@ -1,6 +1,8 @@
 Changes
 *******
 
+- You can now supply a state-transformation function.
+
 0.1.0 (2017-01-05)
 ==================